commit & push dev-exception

+++ src/main/java/com/app/microapplication/controller/ProductServiceController.java
@@ -14,6 +14,7 @@ import org.springframework.http.ResponseEntity;
 
 
 import com.app.microapplication.model.Product;
+import com.app.microapplication.exception.ProductNotFoundException;
 
 @RestController
 public class ProductServiceController {
@@ -34,6 +35,7 @@ public class ProductServiceController {
       return new ResponseEntity<>(productRepo.values(), HttpStatus.OK);
    }
    
+   
    @RequestMapping(value = "/products", method = RequestMethod.POST)
    public ResponseEntity<Object> createProduct(@RequestBody Product product) {
       productRepo.put(product.getId(), product);
@@ -41,7 +43,8 @@ public class ProductServiceController {
    }
    
    @RequestMapping(value = "/products/{id}", method = RequestMethod.PUT)
-   public ResponseEntity<Object> updateProduct(@PathVariable("id") String id, @RequestBody Product product) { 
+   public ResponseEntity<Object> updateProduct(@PathVariable("id") String id, @RequestBody Product product) {
+	  if(!productRepo.containsKey(id)) throw new ProductNotFoundException();
       productRepo.remove(id);
       product.setId(id);
       productRepo.put(id, product);
